Add set_impact_side(), called after dda loop
When we have our ray impact we can use this to set the strike side.
There is a bit of redundancy? confusion? with what was player.side and I
turned into ray.axis so I have added (for now! at least) ray.impact_side
and kept the axis setting. (But moved it so we only set it once,
because why flick it backwards and forwards)