Change axis criteria in set_impact_side
Now matches the N-S / E-W division and map.cub, for example, no longer
shows N at the end of the wall, incorrectly.
BUT this exposes a (nother? new?) problem - correct textures chosen
but they are mirrored.
So do we really *want* this correction? What is the missing part of it?