Fix leak when closing with window controls
ESC was hooked into exit game and cleared the memory properly, but the
window control close button was not, and so we had leaks.

Added a call to exit_game() in main so that if the mlx_loop() ends
without triggering an exit then we still clean up on the way out.