FIX for find_next_line() leaks
If there is a validation failure before the map part of the file,
before, we exited and left the file partially read. This lead to
pointers hanging around in a "still reachable" error reported by
valgrind which some evaluators treat as a problem.

Now when we bail out early, we read the remainder of the file so that we
know it is freed for sure. This seems like it could be abused, e.g. if
someone hooked up an infinitely large file, but that seems less likely
to happen than that an evaluator objects to the still reachable message.

This way we get a lovely "all blocks freed" message and a warm feeling.